we_are_social - Copy - Copy/requirements.txt
arrow==0.7.0
Django==1.9.6
django-forms-bootstrap==3.0.1
django-paypal==0.2.7
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
stripe==1.33.0
